commit
I copy and pasted from the other python project because I was having trouble submitting it otherwise. If you want me to try and send the original I will, just for the purpose of getting it in this is my quickest solution. I will need help in class though. I was having errors dragging it I think due to lack of cloud space.